Enhance the app to see multiple mentions in one message
Enabled the code to process multiple occurances of the ++/-- syntax in
one message iteratively thus handling score updates for multiple
recipients in one message. Changed the handler for selfPlus to be
processed within the handlePlusMinus handler an simply not increment the
score while providing the disapproving random message to the instigator.